split components
