+++ packages/app-photo-loader/photo-loader.tsx
@@ -2,65 +2,15 @@ import React, { useState, useEffect, useRef, useCallback } from "react";
 import styled from "@emotion/styled";
 import Masonry, { ResponsiveMasonry } from "react-responsive-masonry";
 import { SearchInput } from "@ui/core/search";
-import Axios from "axios";
-
+import * as api from "./client";
+import { LoadableGraphicItem } from "./image-item";
+import { EK_CREATE_IMAGE } from "@core/constant/ek.constant";
 ///
 ///  TODO:
 ///  1. Move api client somewhere else
 ///  2. Support for DND
 ///  3. Support for replace fill(s)
 ///
-
-const client = Axios.create({
-  baseURL: "/api",
-});
-
-interface SearchQuery {
-  q: string;
-}
-
-interface GenerativeResponse {
-  q: string;
-  n: number;
-  images: string[];
-  size: "1024x1024" | "512x512" | "256x256";
-}
-
-interface ResourceResponse {
-  q: string;
-  n: number;
-  images: {
-    id: string;
-    thumbnail: string;
-    url: string;
-    raw: string;
-    alt: string;
-    color: string;
-    width: number;
-    height: number;
-    author: {
-      username: string;
-      name: string;
-    };
-  }[];
-  pages: number;
-}
-
-async function fromGenerative(p: SearchQuery): Promise<GenerativeResponse> {
-  const { data } = await client.get("/generative/images", {
-    params: p,
-  });
-
-  return data;
-}
-
-async function fromResources(p: SearchQuery): Promise<ResourceResponse> {
-  const { data } = await client.get("/resources/images", {
-    params: p,
-  });
-
-  return data;
-}
 
 interface PlacableImage {
   thumbnail: string;
@@ -79,9 +29,11 @@ export function PhotoLoader() {
 
   const promptGeneration = async () => {
     setLocked(true);
-    const { images: gens, n } = await fromGenerative({
-      q: query,
-    }).finally(() => setLocked(false));
+    const { images: gens, n } = await api
+      .fromGenerative({
+        q: query,
+      })
+      .finally(() => setLocked(false));
 
     setImages({
       images: gens.map((g) => ({
@@ -107,9 +59,11 @@ export function PhotoLoader() {
       }
 
       setLocked(true);
-      const { images: gens, n } = await fromResources({
-        q: query,
-      }).finally(() => setLocked(false));
+      const { images: gens, n } = await api
+        .fromResources({
+          q: query,
+        })
+        .finally(() => setLocked(false));
 
       if (!isQuerySufficientForGeneration) {
         setImages({
@@ -198,100 +152,6 @@ const GenerateButton = styled.button`
   border-radius: 8px;
   cursor: pointer;
   transition: all 0.2s ease-in-out;
-`;
-
-interface OverlayStyle extends React.CSSProperties {
-  "--progress"?: number;
-}
-
-function LoadableGraphicItem({
-  onResourceReady,
-  name,
-  src,
-}: {
-  onResourceReady?: () => void;
-  name: string;
-  src: string;
-}) {
-  const [isDownloading, setIsDownloading] = useState(false);
-  const [progress, setProgress] = useState(0);
-
-  // TODO: handle on ready (onResourceReady)
-
-  const onclick = () => {
-    setIsDownloading(true);
-
-    // mock progress
-    const interval = setInterval(() => {
-      setProgress((p) => {
-        if (p >= 1) {
-          clearInterval(interval);
-          setIsDownloading(false);
-          return 0;
-        }
-        return p + 0.1;
-      });
-    }, 100);
-  };
-
-  return (
-    <ItemWrapper onClick={onclick}>
-      <img
-        style={{ width: "100%", height: "100%", display: "block" }}
-        src={src}
-        alt={name}
-      />
-      <div
-        className={"loading-overlay"}
-        data-state={isDownloading ? "loading" : "idle"}
-        style={{ "--progress": progress } as OverlayStyle}
-      />
-    </ItemWrapper>
-  );
-}
-
-const ItemWrapper = styled.div`
-  position: relative;
-  /*  */
-
-  .loading-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    opacity: 1;
-
-    &[data-state="idle"] {
-      opacity: 0;
-    }
-
-    &[data-state="loading"] {
-      opacity: 1;
-    }
-
-    &[data-state="idle"]::before {
-      transform-origin: right;
-    }
-
-    &[data-state="loading"]::before {
-      transform-origin: left;
-    }
-
-    transition: opacity 0.5s ease-in-out;
-  }
-
-  .loading-overlay:before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    background-color: rgba(255, 255, 255, 0.2);
-    transform: scaleX(var(--progress));
-    transition: transform 0.3s linear;
-  }
 `;
 
 /**
